% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_2021_merged.R
\docType{data}
\name{ab_2021_merge}
\alias{ab_2021_merge}
\title{AmericasBarometer Merge 2021}
\format{
An object of class \code{"data.frame"}
}
\source{
\href{https://www.vanderbilt.edu/lapop/raw-data.php}{LAPOP Data Sets}
}
\usage{
data(ab_2021_merge)
}
\description{
Data from the AmericasBarometer survey 2021 wave, with data merged from all countries
}
\examples{

data(ab_2021_merge)
head(ab_2021_merge)


}
\references{
The AmericasBarometer by the LAPOP Lab
(\href{www.vanderbilt.edu/lapop})
}
\keyword{datasets}
