% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapop_fonts.R
\name{lapop_fonts}
\alias{lapop_fonts}
\title{LAPOP Fonts}
\usage{
lapop_fonts()
}
\description{
This function loads fonts needed for LAPOP graph formatting.
No arguments needed; just run lapop_fonts() at the beginning of your session.
}
\author{
Luke Plutowski, \email{luke.plutowski@vanderbilt.edu}
}
