% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapop_stacked.R
\name{lapop_sb}
\alias{lapop_sb}
\title{LAPOP Bar Graphs}
\usage{
lapop_sb(
  data,
  outcome_var = data$prop,
  prop_labels = data$proplabel,
  var_labels = data$varlabel,
  value_labels = data$vallabel,
  lang = "en",
  main_title = "",
  subtitle = "",
  source_info = "",
  rev_values = FALSE,
  rev_variables = FALSE,
  subtitle_h_just = 0,
  color_scheme = c("#2D708E", "#1F9689", "#00ADA9", "#21A356", "#568424", "#ACB014")
)
}
\arguments{
\item{data}{Data Frame. Dataset to be used for analysis.  The data frame should have columns
titled varlabel (name(s)/label(s) of variable(s) of interest; character), vallabel (names/labels of values for each variable; character),
prop (outcome variable value; numeric), and proplabel (text of outcome variable value; character).
Default: None (must be provided).}

\item{outcome_var, prop_labels, var_labels, value_labels}{Numeric, character, character, character.
Each component of the data to be plotted can be manually specified in case
the default columns in the data frame should not be used (if, for example, the values for a given
variable were altered and stored in a new column).}

\item{lang}{Character.  Changes default subtitle text and source info to either Spanish or English.
Will not translate input text, such as main title or variable labels.  Takes either "en" (English)
or "es" (Spanish).  Default: "en".}

\item{main_title}{Character.  Title of graph.  Default: None.}

\item{subtitle}{Character.  Describes the values/data shown in the graph, e.g., "Percent who support...".
Default: None.}

\item{source_info}{Character.  Information on dataset used (country, years, version, etc.),
which is added to the end of "Source: AmericasBarometer" in the bottom-left corner of the graph.
Default: None (only "Source: AmericasBarometer" will be printed).}

\item{rev_values}{Logical.  Should the order of the values for each variable be reversed?  Default: FALSE.}

\item{rev_variables}{Logical.  Should the order of the variables be reversed?  Default: FALSE.}

\item{subtitle_h_just}{Numeric.  Move the subtitle/legend text left (negative numbers) or right (positive numbers).
Ranges from -100 to 100.  Default: 0.}

\item{color_scheme}{Character.  Color of data bars for each value.  Allows up to 6 values.
Takes hex numbers, beginning with "#".
Default: c("#2D708E", "#1F9689", "#00ADA9", "#21A356", "#568424", "#ACB014")
(navy blue, turquoise, teal, green, sap green, pea soup).}
}
\value{
Returns a ggplot graph.
}
\description{
This function shows a stacked bar graph using LAPOP formatting.
}
\examples{

df <- data.frame(varlabel = c(rep("Politicians can\nidentify voters", 5),
                              rep("Wealthy can\nbuy results", 5),
                              rep("Votes are\ncounted correctly", 5)),
                 vallabel = rep(c("Always", "Often", "Sometimes",
                                  "Never", "Other"), 3),
                 prop = c(36, 10, 19, 25, 10, 46, 10, 23, 11, 10, 35,
                          10, 32, 13, 10),
                 proplabel = c("36\%", "10\%", "19\%", "25\%", "10\%", "46\%",
                               "10\%", "23\%", "11\%", "10\%", "35\%", "10\%",
                               "32\%", "13\%", "10\%"))

lapop_sb(df,
         main_title = "Trust in key features of the electoral process is low in Latin America",
         subtitle = "\% believing it happens:",
         source_info = "2019")

}
\author{
Luke Plutowski, \email{luke.plutowski@vanderbilt.edu}
}
