% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapop_tsmulti.R
\name{lapop_tsmulti}
\alias{lapop_tsmulti}
\title{LAPOP Multi-line Time-Series Graphs}
\usage{
lapop_tsmulti(
  data,
  outcome_var = data$prop,
  lower_bound = data$lb,
  upper_bound = data$ub,
  wave_var = as.character(data$wave),
  label_var = data$proplabel,
  point_var = data$prop,
  varlabel = data$varlabel,
  ymin = plyr::round_any(min(outcome_var, na.rm = TRUE) - 5, 5, f = floor),
  ymax = plyr::round_any(max(outcome_var, na.rm = TRUE) + 5, 5, f = ceiling),
  main_title = "",
  source_info = "",
  subtitle = "",
  lang = "en",
  subtitle_h_just = 40,
  subtitle_v_just = -20,
  color_scheme = c("#7030A0", "#3CBC70", "#1F968B", "#95D840")
)
}
\arguments{
\item{data}{Data Frame. Dataset to be used for analysis.  The data frame should have columns
titled wave (survey wave/year; character vector), prop (outcome variable; numeric),
proplabel (text of outcome variable; character); lb (lower bound of estimate; numeric),
and ub (upper bound of estimate; numeric). Default: None (must be supplied).}
}
\description{
Documentation TBD
}
\author{
Luke Plutowski, \email{luke.plutowski@vanderbilt.edu}
}
