% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpr_dumb.R
\name{lpr_dumb}
\alias{lpr_dumb}
\title{LAPOP Dumbbell Graphs}
\usage{
lpr_dumb(
  data,
  outcome,
  xvar = "pais",
  over,
  rec = c(1, 1),
  ci_level = 0.95,
  mean = FALSE,
  filesave = "",
  cfmt = "",
  sort = "prop2",
  order = "hi-lo",
  ttest = FALSE,
  keep_nr = FALSE
)
}
\arguments{
\item{data}{A survey object.  The data that should be analyzed.}

\item{outcome}{Outcome variable(s) of interest to be plotted across countries
and waves, supplied as a character string or vector of strings.}

\item{xvar}{Character. The grouping variable to be plotted
along the x-axis (technically, the vertical axis for lapop_dumb). Usually
country (pais). Default: "pais".}

\item{over}{Numeric. A vector of values for "wave" that specify which two
waves should be included in the plot.}

\item{rec}{Numeric. The minimum and maximum values of the outcome variable that
should be included in the numerator of the percentage.  For example, if the variable
is on a 1-7 scale and rec is c(5, 7), the function will show the percentage who chose
an answer of 5, 6, 7 out of all valid answers.  Can also supply one value only,
to produce the percentage that chose that value out of all other values.
Default: c(1, 1).}

\item{ci_level}{Numeric. Confidence interval level for estimates.  Default: 0.95}

\item{mean}{Logical.  If TRUE, will produce the mean of the variable rather than
recoding to percentage.  Default: FALSE.}

\item{filesave}{Character.  Path and file name to save the dataframe as csv.}

\item{cfmt}{Character. Changes the format of the numbers displayed above the bars.
Uses sprintf string formatting syntax. Default is whole numbers for percentages
and tenths place for means.}

\item{sort}{Character. On what value the bars are sorted.
Options are "prop1" (for the value of the outcome variable in wave 1), "prop2"
(default; for the value of the outcome variable in wave 2), "xv" (for
the underlying values of the x variable), "xl" (for the labels of the x variable,
i.e., alphabetical), and "diff" (for the difference between the outcome between
the two waves).}

\item{order}{Character.  How the bars should be sorted.  Options are "hi-lo"
(default) or "lo-hi".}

\item{ttest}{Logical.  If TRUE, will conduct pairwise t-tests for difference
of means between all pais-wave combinations and save them in attr(x,
"t_test_results"). Default: FALSE.}

\item{keep_nr}{Logical.  If TRUE, will convert "don't know" (missing code .a)
and "no response" (missing code .b) into valid data (value = 99) and use them
in the denominator when calculating percentages.  The default is to examine
valid responses only.  Default: FALSE.}
}
\value{
Returns a data frame, with data formatted for visualization by lapop_dumb()
}
\description{
This function creates dataframes which can then be input in lapop_dumb for
comparing means of a variable across countries and two waves using LAPOP formatting.
}
\examples{

require(lapop); data(cm23)

# Set Survey Context
cm23lpr <- lpr_data(cm23)

# Single outcome over years
lpr_dumb(cm23lpr,
outcome = "ing4",
rec = c(5, 7),
over = c("2018/19", "2023"),
sort = "diff")

# Multiple outcomes over years
lpr_dumb(cm23lpr,
outcome=c("b13","b21", "b31"),
rec=c(5,7),
over=c("2018/19", "2023"))

}
\author{
Luke Plutowski, \email{luke.plutowski@vanderbilt.edu} & Robert Vidigal, \email{robert.vidigal@vanderbilt.edu}
}
