% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpr_mline_mv.R
\name{lpr_mline_mv}
\alias{lpr_mline_mv}
\title{LAPOP Multi-Line, Multi-Variable Time Series Graph Pre-Processing}
\usage{
lpr_mline_mv(
  data,
  outcome,
  rec1 = c(1, 1),
  rec2 = c(1, 1),
  rec3 = c(1, 1),
  rec4 = c(1, 1),
  use_wave = FALSE,
  ci_level = 0.95,
  mean = FALSE,
  filesave = "",
  cfmt = "",
  ttest = FALSE,
  keep_nr = FALSE
)
}
\arguments{
\item{data}{A survey object.  The data that should be analyzed.}

\item{outcome}{Character vector.  Outcome variables of interest to be plotted
across time.}

\item{rec1, rec2, rec3, rec4}{Numeric. The minimum and maximum values of the outcome
variable that should be included in the numerator of the percentage.
For example, if the variable is on a 1-7 scale and rec is c(5, 7), the
function will show the percentage who chose an answer of 5, 6, 7 out of
all valid answers.  Can also supply one value only, to produce the percentage
that chose that value out of all other values. Default: c(1, 1).}

\item{use_wave}{Logical.  If TRUE, will use "wave" for the x-axis; otherwise,
will use "year".  Default: FALSE.}

\item{ci_level}{Numeric. Confidence interval level for estimates.  Default: 0.95}

\item{mean}{Logical.  If TRUE, will produce the mean of the variable rather than
rescaling to percentage.  Default: FALSE.}

\item{filesave}{Character.  Path and file name to save the dataframe as csv.}

\item{cfmt}{Character. changes the format of the numbers displayed above the bars.
Uses sprintf string formatting syntax. Default is whole numbers for percentages
and tenths place for means.}

\item{ttest}{Logical.  If TRUE, will conduct pairwise t-tests for difference
of means between all individual x levels and save them in attr(x,
"t_test_results"). Default: FALSE.}

\item{keep_nr}{Logical.  If TRUE, will convert "don't know" (missing code .a)
and "no response" (missing code .b) into valid data (value = 99) and use them
in the denominator when calculating percentages.  The default is to examine
valid responses only.  Default: FALSE.}
}
\value{
Returns a data frame, with data formatted for visualization by lapop_mline
}
\description{
This function creates a dataframe which can then be input in lapop_mline. It
will show the values of between 1 and 4 variables across time (essentially,
it allows you to do lpr_ts for multiple variables).  The lpr_mline function
instead shows the value of one outcome variable, broken down by values of
a secondary variable.
}
\examples{

\dontrun{lpr_mline_mv(ttots,
outcome = c("ing4", "pn4", "b2", "vic1ext"),
rec1 = c(5, 7),
rec2 = c(1, 2),
rec3 = c(5, 7),
rec4 = c(1, 1),
use_wave = TRUE)}

}
\author{
Luke Plutowski, \email{luke.plutowski@vanderbilt.edu}
}
