% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpr_stack_over.R
\name{lpr_stack_over}
\alias{lpr_stack_over}
\title{LAPOP Stacked-Over Bar Graph Pre-Processing}
\usage{
lpr_stack_over(
  data,
  outcome,
  over,
  sort = "xv",
  order = "lo-hi",
  filesave = "",
  keep_nr = FALSE
)
}
\arguments{
\item{data}{The data that should be analyzed. It requires a survey object
from lpr_data() function.}

\item{outcome}{Character. Outcome variable to be plotted.}

\item{over}{Character. Grouping variable.  The function breaks down the outcome
variable over distinct values of the "over" variable.}

\item{sort}{Character. On what value the bars are sorted. Options are "xv"
(default; sort on the underlying values of the value labels), "xl" (on the value
labels themselves, i.e. alphabetically), or "y" (on the proportions of the
outcome variable(s)).}

\item{order}{Character. How the bars should be sorted. Options are "lo-hi"
(default) or "hi-lo".}

\item{filesave}{Character. Path and file name to save the dataframe as csv.}

\item{keep_nr}{Logical. If TRUE, will convert "don't know" (missing code .a)
and "no response" (missing code .b) into valid data (value = 99) and use them
in the denominator when calculating percentages.  The default is to examine
valid responses only.  Default: FALSE.}
}
\value{
Returns a data frame, with data formatted for visualization by lapop_stack
}
\description{
Like lpr_stack(), this function creates dataframes which can then be input
in lapop_stack() for plotting variables categories with a stacked bar graph.
Unlike lpr_stack(), which shows a bar for each specified outcome variable,
this function takes one outcome variable, and plots it across all values
of a secondary "over" variable.
}
\examples{

\dontrun{lapop_stack(data = gm, outcome = "aoj11", over = "wave"))}

}
\author{
Luke Plutowski, \email{luke.plutowski@vanderbilt.edu}
}
