% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_helper.R
\name{stack_helper}
\alias{stack_helper}
\title{Stacked Bar Graph Pre-Processing}
\usage{
stack_helper(
  data,
  outcome,
  by = NULL,
  sort = "xv",
  order = "lo-hi",
  filesave = "",
  keep_nr = FALSE
)
}
\arguments{
\item{data}{The data that should be analyzed.}

\item{outcome}{Character vector. Vector of variable(s) be plotted.}

\item{by}{Character. Grouping variable.  If by is provided, function will break
down the outcome variable by distinct values of the "by" variable (instead of
showing distribution of multiple outcome variables).}

\item{sort}{Character. On what value the bars are sorted. Options are "xv"
(default; sort on the underlying values of the value labels), "xl" (on the value
labels themselves, i.e. alphabetically), or "y" (on the proportions of the
outcome variable(s)).}

\item{order}{Character. How the bars should be sorted. Options are "lo-hi"
(default) or "hi-lo".}

\item{filesave}{Character. Path and file name to save the dataframe as csv.}

\item{keep_nr}{Logical. If TRUE, will convert "don't know" (missing code .a)
and "no response" (missing code .b) into valid data (value = 99) and use them
in the denominator when calculating percentages.  The default is to examine
valid responses only.  Default: FALSE.}
}
\value{
Returns a data frame, with data formatted for visualization by lapop_stack
}
\description{
This function creates dataframes which can then be used to create stacked
bar plots.
If "by" is specified, the function shows the distribution of the outcome variable
by distinct values of a secondary variable.  If "by" is not provided, it will show
the distribution of all specified outcome variables.
}
\examples{

\dontrun{lapop_stack(data = gm, outcome = c("countfair1", "countfair3"))}

}
\author{
Luke Plutowski & Robert Vidigal
}
