% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world.R
\docType{data}
\name{world}
\alias{world}
\title{bra23: Single-country Single-year Dataset}
\format{
A data frame
\describe{
\item{pais}{Country name}
\item{pais_lab}{Country ISO2 code}
\item{geometry}{Polygon with geometry}
}
}
\source{
LAPOP AmericasBarometer (https://www.vanderbilt.edu/lapop/)
}
\usage{
world
}
\description{
A dataset containing the World Map geometry for lapop_map() function.
}
\keyword{datasets}
