% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ym23.R
\docType{data}
\name{ym23}
\alias{ym23}
\title{ym23: Multi-country Single-year Dataset}
\format{
A data frame
\describe{
\item{b12}{Trust in Armed Forces}
\item{b18}{Trust in National Police}
\item{ing4}{Support for Democracy}
\item{pn4}{Satisfaction with Democracy}
\item{vb21n}{Influence Political Change}
\item{q14f}{Migration Intentions}
\item{edre}{Education}
\item{wealth}{Wealth}
\item{q1tc_r}{Gender}
\item{upm}{Primary Sampling Unit}
\item{strata}{Stratification}
\item{wave}{Survey round year for regional or multi-country data}
\item{pais}{Country of survey}
\item{year}{Survey round year for single-country data}
\item{weight1500}{Cross-country and cross-time weight}
}
}
\source{
LAPOP AmericasBarometer (https://www.vanderbilt.edu/lapop/)  # Ensure this line has a valid source.
}
\usage{
ym23
}
\description{
A dataset containing the AmericasBarometer Year Merge of 2023.
}
\keyword{datasets}
