% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm23.R
\docType{data}
\name{cm23}
\alias{cm23}
\title{cm23 Dataset}
\format{
A data frame
\describe{
  \item{aoj11}{Description of Column}
  \item{ing4}{Description of Column}
  \item{b13}{Description of Column}
  \item{b18}{Description of Column}
  \item{b21}{Description of Column}
  \item{b31}{Description of Column}
  \item{b12}{Description of Column}
  \item{wave}{Description of Column}
  \item{pais}{Description of Column}
  \item{year}{Description of Column}
  \item{b18}{Description of Column}
  ...
}
}
\source{
LAPOP AmericasBarometer (https://www.vanderbilt.edu/lapop/)
}
\usage{
cm23
}
\description{
A dataset containing the AmericasBarometer Country Merge up to 2023 (Brazil).
}
\keyword{datasets}
