% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapop_fonts_design.R
\name{lapop_fonts_design}
\alias{lapop_fonts_design}
\title{LAPOP Fonts (design)}
\usage{
lapop_fonts_design()
}
\value{
No return value, called for side effects
}
\description{
This function loads fonts needed for LAPOP graph formatting.  In contrast to lapop_fonts(),
this renders text as text instead of polygons, which allows post-hoc editing.
}
\examples{
\dontrun{lapop_fonts_design()}

}
\author{
Luke Plutowski, \email{luke.plutowski@vanderbilt.edu}
}
