% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapop_map.R
\name{lapop_map}
\alias{lapop_map}
\title{LAPOP Map Graph}
\usage{
lapop_map(
  data,
  outcome = "value",
  pais_lab = "pais_lab",
  survey = c("CSES", "AmericasBarometer"),
  zoom = 1,
  title = NULL,
  palette = c("#F2A344", "#D97A1E", "#BF5A00", "#8A3900", "#4A1E00")
)
}
\arguments{
\item{data}{A data frame containing ISO2 country codes and a value to map.}

\item{outcome}{String. Column name containing the numeric or categorical variable to visualize.}

\item{pais_lab}{String. Column name containing ISO2 country codes (e.g., `"US"`, `"BR"`).}

\item{survey}{Either `"CSES"` (full world map) or `"AmericasBarometer"` (Americas only).}

\item{zoom}{Numeric (0–1). Controls how tightly the map zooms when `survey = "AmericasBarometer"`. Default is `1`.}

\item{title}{Optional plot title.}

\item{palette}{Vector of up to 5 colors for continuous and factor variables.}
}
\value{
A `ggplot2` choropleth map object.
}
\description{
`lapop_map()` generates a stylized choropleth map using ISO2 country codes
from both continuous and factor variables. It is designed to map cross-country
results from `lpr_cc()` and supports either a full world map (`survey = "CSES"`)
# or an Americas-only map (`survey = "AmericasBarometer"`).
}
\examples{
\dontrun{
# Continuous variable example
data_cont <- data.frame(
  vallabel = c("US", "AR", "VE", "CH", "EC"),
  prop = c(37, 52, 94, 17, 69)
)
lapop_map(data_cont, pais_lab = "vallabel", outcome = "prop",
          survey = "AmericasBarometer", zoom = 0.9)

# Factor variable example
data_fact <- data.frame(
  valalbel = c("CA", "BR", "MX", "PE", "CO"),
  group = c("A","A","B","B","C")
)
lapop_map(data_fact, pais_lab = "vallabel", outcome = "group",
          survey = "AmericasBarometer")
}

}
\author{
Robert Vidigal, \email{robert.vidigal@vanderbilt.edu}
}
