% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpr_ccm.R
\name{lpr_ccm}
\alias{lpr_ccm}
\title{LAPOP Grouped Bar Graph Pre-Processing}
\usage{
lpr_ccm(
  data,
  outcome_vars,
  xvar = "pais_lab",
  rec1 = c(1, 1),
  rec2 = c(1, 1),
  rec3 = c(1, 1),
  ci_level = 0.95,
  mean = FALSE,
  filesave = "",
  cfmt = "",
  sort = "y",
  order = "hi-lo",
  ttest = FALSE,
  keep_nr = FALSE
)
}
\arguments{
\item{data}{A survey object.  The data that should be analyzed.}

\item{outcome_vars}{Character vector.  Outcome variable(s) of interest to be plotted
across country (or other x variable). Max of 3 variables.}

\item{xvar}{Character string. Outcome variables are broken down by this variable.
Default: pais_lab.}

\item{rec1, rec2, rec3}{Numeric. The minimum and maximum values of the outcome variable that
should be included in the numerator of the percentage.  For example, if the variable
is on a 1-7 scale and rec1 is c(5, 7), the function will show the percentage who chose
an answer of 5, 6, 7 out of all valid answers.  Can also supply one value only,
to produce the percentage that chose that value out of all other values.
Default: c(1, 1).}

\item{ci_level}{Numeric. Confidence interval level for estimates.  Default: 0.95}

\item{mean}{Logical.  If TRUE, will produce the mean of the variable rather than
rescaling to percentage.  Default: FALSE.}

\item{filesave}{Character.  Path and file name to save the dataframe as csv.}

\item{cfmt}{Character. Changes the format of the numbers displayed above the bars.
Uses sprintf string formatting syntax. Default is whole numbers for percentages
and tenths place for means.}

\item{sort}{Character. On what value the bars are sorted.
Options are "y" (default; for the value of the first outcome variable), "xv" (for
the underlying values of the x variable), "xl" (for the labels of the x variable,
i.e., alphabetical).}

\item{order}{Character.  How the bars should be sorted.  Options are "hi-lo"
(default) or "lo-hi".}

\item{ttest}{Logical.  If TRUE, will conduct pairwise t-tests for difference
of means between all outcomes vs. all x-vars and save them in attr(x,
"t_test_results"). Default: FALSE.}

\item{keep_nr}{Logical.  If TRUE, will convert "don't know" (missing code .a)
and "no response" (missing code .b) into valid data (value = 99) and use them
in the denominator when calculating percentages.  The default is to examine
valid responses only.  Default: FALSE.}
}
\value{
Returns a data frame, with data formatted for visualization by lapop_ccm
}
\description{
This function creates dataframes which can then be input in lapop_ccm for
comparing values for multiple variables across countries with a bar graph
using LAPOP formatting.
}
\examples{

\dontrun{lpr_ccm(gm23,
outcome_vars = c("vic1ext", "aoj11"),
rec1 = c(1, 1),
rec2 = c(3, 4),
ttest = TRUE)}

}
\author{
Luke Plutowski, \email{luke.plutowski@vanderbilt.edu}
}
