% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpr_coef.R
\name{lpr_coef}
\alias{lpr_coef}
\title{LAPOP Regression Coefficients Graph Pre-Processing}
\usage{
lpr_coef(
  outcome = NULL,
  xvar = NULL,
  model = "gaussian",
  lprdata = NULL,
  estimate = c("coef"),
  vlabs = NULL,
  omit = NULL,
  filesave = NULL,
  replace = FALSE,
  level = 0.95
)
}
\arguments{
\item{outcome}{Dependent variable for the svyglm regression model. (e.g., "outcome_name"). Only one variable allowed.}

\item{xvar}{Vector of independent variables for the svyglm regression model (e.g., "xvar1+xvar2+xvar3" and so on). Multiple variables are allowed.}

\item{model}{Model family object for glm. Default is linear regression (i.e., "gaussian"). For a logit model, use model="binomial"}

\item{lprdata}{Survey design data from lpr_data() output.}

\item{estimate}{Character. Graph either the coefficients (i.e., `coef`) or the change in probabilities (i.e., `contrast`). Default is "coef."}

\item{vlabs}{Character. Rename variable labels to be displayed in the graph produced by lapop_coef(). For instance, vlabs=c("old_varname" = "new_varname").}

\item{omit}{Character. Do not display coefficients for these independent variables. Default is to display all variables included in the model. To omit any variables you need to include the raw "varname" in the omit argument.}

\item{filesave}{Character. Path and file name with csv extension to save the dataframe output.}

\item{replace}{Logical. Replace the dataset output if it already exists. Default is FALSE.}

\item{level}{Numeric. Set confidence level in numeric values; default is 95 percent.}
}
\value{
Returns a data frame, with data formatted for visualization by lapop_coef
}
\description{
This function creates a data frame which can then be input in lapop_coef() for
plotting regression coefficients graph using LAPOP formatting.
}
\examples{

\dontrun{dataLAPOP<-lpr_data(dataset)}
\dontrun{svyglm_object<-survey::svyglm(formula, desig, family)}
\dontrun{Example 1: svyglm_linear<-survey::svyglm(fs2~it1+idio2+edad, data=dataLAPOP, family="gaussian")}
\dontrun{lpr_coef(outcome="fs2", xvar="it1+idio2+edad", lprdata=dataLAPOP, est="coef")}
\dontrun{Example 2: svyglm_logit<-survey::svyglm(fs2~it1+idio2+edad, data=dataLAPOP, family="binomial")}
\dontrun{lpr_coef(outcome="fs2", xvar="it1+idio2+edad", lprdata=dataLAPOP, model="binomial", est="contrast")}

}
\author{
Robert Vidigal, \email{robert.vidigal@vanderbilt.edu}
}
