% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpr_data.R
\name{lpr_data}
\alias{lpr_data}
\title{LAPOP Data Processing}
\usage{
lpr_data(data_path, wt = FALSE)
}
\arguments{
\item{data_path}{A dataframe of LAPOP survey data.}

\item{wt}{Logical.  If TRUE, use wt instead of weight1500.  Default: FALSE.}
}
\value{
Returns a svy_tbl object
}
\description{
This function takes LAPOP datasets and adds survey features, outputting a
svy_tbl object that can then be analyzed using lpr_ commands.
}
\examples{
# Single-country (wt)
bra23w <- lpr_data(bra23, wt = TRUE)
print(bra23w)

# Single-country multi-year (weight1500)
cm23w <- lpr_data(cm23)
print(cm23w)

# Multi-country single-year (weight1500)
ym23w <- lpr_data(cm23)
print(cm23w)
}
\author{
Luke Plutowski, \email{luke.plutowski@vanderbilt.edu} & Robert Vidigal, \email{robert.vidigal@vanderbilt.edu}
}
