% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpr_data.R
\name{lpr_data}
\alias{lpr_data}
\title{LAPOP Data Processing}
\usage{
lpr_data(data_path, wt = FALSE)
}
\arguments{
\item{data_path}{A dataframe of LAPOP survey data.}

\item{wt}{Logical.  If TRUE, use wt instead of weight1500.  Default: FALSE.}
}
\value{
Returns a svy_tbl object
}
\description{
This function takes LAPOP datasets and adds survey features, outputting a
svy_tbl object that can then be analyzed using lpr_ commands.
}
\examples{
\dontrun{
gm23 <- lpr_data("Merge 2023 LAPOP AmericasBarometer (v1.0s).dta")}
}
\author{
Luke Plutowski, \email{luke.plutowski@vanderbilt.edu}
}
