% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpr_data.R
\name{lpr_data}
\alias{lpr_data}
\title{LAPOP Data Processing}
\usage{
lpr_data(data_path)
}
\arguments{
\item{data}{A dataframe of LAPOP survey data.}
}
\value{
Returns a svy_tbl object
}
\description{
This function takes LAPOP datasets and adds survey features, outputting a
svy_tbl object that can then be analyzed using lpr_ commands.
}
\examples{

\dontrun{
gm23 <- haven::read_dta("Merge 2023 LAPOP AmericasBarometer (v1.0s).dta")

gm23s <- lpr_data(gm23)
}

}
\author{
Luke Plutowski, \email{luke.plutowski@vanderbilt.edu}
}
