% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpr_mline.R
\name{lpr_mline}
\alias{lpr_mline}
\title{LAPOP Multi-Line Time Series Graphs}
\usage{
lpr_mline(
  data,
  outcome,
  xvar,
  rec = c(1, 1),
  use_wave = FALSE,
  ci_level = 0.95,
  mean = FALSE,
  filesave = "",
  cfmt = "",
  ttest = FALSE,
  keep_nr = FALSE
)
}
\arguments{
\item{data}{A survey object.  The data that should be analyzed.}

\item{outcome}{Outcome variable of interest to be plotted across countries.}

\item{xvar}{Variable on which to separate out the outcome variable.  In other words,
the line graph will produce multiple lines for each value of xvar (technically, it is
the z-variable, not the x variable, which is year/wave).}

\item{rec}{Numeric. The minimum and maximum values of the outcome variable that
should be included in the numerator of the percentage.  For example, if the variable
is on a 1-7 scale and rec is c(5, 7), the function will show the percentage who chose
an answer of 5, 6, 7 out of all valid answers.  Default: c(1, 1).}

\item{use_wave}{Logical.  If TRUE, will use "wave" for the x-axis; otherwise,
will use "year".  Default: FALSE.}

\item{ci_level}{Numeric. Confidence interval level for estimates.  Default: 0.95}

\item{mean}{Logical.  If TRUE, will produce the mean of the variable rather than
recoding to percentage.  Default: FALSE.}

\item{filesave}{Character.  Path and file name to save the dataframe as csv.}

\item{cfmt}{changes the format of the numbers displayed above the bars.
Uses sprintf string formatting syntax. Default is whole numbers for percentages
and tenths place for means.}

\item{ttest}{Logical.  If TRUE, will conduct pairwise t-tests for difference
of means between all individual year-xvar levels and save them in attr(x,
"t_test_results"). Default: FALSE.}

\item{keep_nr}{Logical.  If TRUE, will convert "don't know" (missing code .a)
and "no response" (missing code .b) into valid data (value = 99) and use them
in the denominator when calculating percentages.  The default is to examine
valid responses only.  Default: FALSE.}
}
\value{
Returns a data frame, with data formatted for visualization by lapop_cc
}
\description{
This function creates dataframes which can then be input in lapop_mline for
comparing values across countries with a bar graph using LAPOP formatting.
}
\examples{

\dontrun{lpr_cc(data = gms, outcome = ing4, num = c(5, 7))}

}
\author{
Luke Plutowski, \email{luke.plutowski@vanderbilt.edu}
}
