% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpr_mover.R
\name{lpr_mover}
\alias{lpr_mover}
\title{LAPOP "Multiple-Over" Breakdown Graphs}
\usage{
lpr_mover(
  data,
  outcome,
  grouping_vars,
  rec = c(1, 1),
  ci_level = 0.95,
  mean = FALSE,
  filesave = "",
  cfmt = "",
  ttest = FALSE,
  keep_nr = FALSE
)
}
\arguments{
\item{data}{A survey object.  The data that should be analyzed.}

\item{outcome}{Character. Outcome variable of interest to be plotted across secondary
variable(s).}

\item{grouping_vars}{A character vector specifying one or more grouping variables.
For each variable, the function calculates the average of the outcome variable,
broken down by the distinct values within the grouping variable(s).}

\item{rec}{Numeric. The minimum and maximum values of the outcome variable that
should be included in the numerator of the percentage.  For example, if the variable
is on a 1-7 scale and rec is c(5, 7), the function will show the percentage who chose
an answer of 5, 6, 7 out of all valid answers.  Can also supply one value only,
to produce the percentage that chose that value out of all other values.
Default: c(1, 1).}

\item{ci_level}{Numeric. Confidence interval level for estimates.  Default: 0.95}

\item{mean}{Logical.  If TRUE, will produce the mean of the variable rather than
recoding to percentage.  Default: FALSE.}

\item{filesave}{Character.  Path and file name to save the dataframe as csv.}

\item{cfmt}{Changes the format of the numbers displayed above the bars.
Uses sprintf string formatting syntax. Default is whole numbers for percentages
and tenths place for means.}

\item{ttest}{Logical.  If TRUE, will conduct pairwise t-tests for difference
of means between all individual year-xvar levels and save them in attr(x,
"t_test_results"). Default: FALSE.}

\item{keep_nr}{Logical.  If TRUE, will convert "don't know" (missing code .a)
and "no response" (missing code .b) into valid data (value = 99) and use them
in the denominator when calculating percentages.  The default is to examine
valid responses only.  Default: FALSE.}
}
\value{
Returns a data frame, with data formatted for visualization by lapop_mover
}
\description{
This function creates dataframes which can then be input in lapop_mover for
comparing means across values of secondary variable(s) using LAPOP formatting.
}
\examples{

\dontrun{lpr_mover(data = gm23,
 outcome = "ing4",
 grouping_vars = c("q1tc_r", "edad", "edre", "wealth"),
 num = c(5, 7)}

}
\author{
Luke Plutowski, \email{luke.plutowski@vanderbilt.edu}
}
