% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R, R/ym23.R
\docType{data}
\name{ym23}
\alias{ym23}
\title{ym23 Dataset}
\format{
A data frame
\describe{
  \item{aoj11}{Description of Column}
  \item{ing4}{Description of Column}
  \item{b13}{Description of Column}
  \item{b21}{Description of Column}
  \item{b31}{Description of Column}
  \item{b12}{Description of Column}
  \item{wave}{Description of Column}
  \item{b18}{Description of Column}
  \item{pn4}{Description of Column}
  \item{edre}{Description of Column}
  \item{wealth}{Description of Column}
  \item{q1tc_r}{Description of Column}
  \item{vb21n}{Description of Column}
  \item{q14f}{Description of Column}
}

A data frame
\describe{
  \item{aoj11}{Description of Column}
  \item{ing4}{Description of Column}
  \item{b13}{Description of Column}
  \item{b21}{Description of Column}
  \item{b31}{Description of Column}
  \item{b12}{Description of Column}
  \item{wave}{Description of Column}
  \item{b18}{Description of Column}
  \item{pn4}{Description of Column}
  \item{edre}{Description of Column}
  \item{wealth}{Description of Column}
  \item{q1tc_r}{Description of Column}
  \item{vb21n}{Description of Column}
  \item{q14f}{Description of Column}
}
}
\source{
LAPOP AmericasBarometer (https://www.vanderbilt.edu/lapop/)  # Ensure this line has a valid source.

LAPOP AmericasBarometer (https://www.vanderbilt.edu/lapop/)  # Ensure this line has a valid source.
}
\usage{
ym23

ym23
}
\description{
A dataset containing information about the AmericasBarometer Year Merge up to 2023.

A dataset containing information about the AmericasBarometer Year Merge up to 2023.
}
\keyword{datasets}
