% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ym23.R
\docType{data}
\name{ym23}
\alias{ym23}
\title{Multi-country Single-year AmericasBarometer 2023 Dataset}
\format{
A data frame
\describe{
  \item{b12}{Description of Column}
  \item{b18}{Description of Column}
  \item{wave}{Description of Column}
  \item{pais}{Description of Column}
  \item{year}{Description of Column}
  \item{ing4}{Description of Column}
  \item{pn4}{Description of Column}
  \item{vb21n}{Description of Column}
  \item{q14f}{Description of Column}
  \item{edre}{Description of Column}
  \item{wealth}{Description of Column}
  \item{q1tc_r}{Description of Column}
  \item{upm}{Description of Column}
  \item{strata}{Description of Column}
  \item{weight1500}{Description of Column}
}
}

\source{
LAPOP Lab AmericasBarometer 2023 (https://www.vanderbilt.edu/lapop/)
}

\usage{
ym23
}

\description{
A dataset containing multiple countries from the AmericasBarometer 2023 survey round.
}

\keyword{datasets}
